% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrivingFactors.R
\name{Decomposition}
\alias{Decomposition}
\title{Decomposition of attributed deaths}
\usage{
Decomposition(serie, crf, ci = "MEAN", G, D, P, A, M, L, D_cf = NULL, from, to)
}
\arguments{
\item{serie}{possible cumulative change in input driver series, listed as follows:
1. PG-PA-EXP-ORF
2. PG-PA-ORF-EXP
3. PG-EXP-PA-ORF
4. PG-EXP-ORF-PA
5. PG-ORF-PA-EXP
6. PG-ORF-EXP-PA
7. PA-PG-EXP-ORF
8. PA-PG-ORF-EXP
9. PA-EXP-PG-ORF
10. PA-EXP-ORF-PG
11. PA-ORF-PG-EXP
12. PA-ORF-EXP-PG
13. EXP-PG-PA-ORF
14. EXP-PG-ORF-PA
15. EXP-PA-PG-ORF
16. EXP-PA-ORF-PG
17. EXP-ORF-PA-PG
18. EXP-ORF-PG-PA
19. ORF-PG-PA-EXP
20. ORF-PG-EXP-PA
21. ORF-PA-PG-EXP
22. ORF-PA-EXP-PG
23. ORF-EXP-PG-PA
24. ORF-EXP-PA-PG}

\item{crf}{refers to `CRF` in `Mortality()`}

\item{ci}{refers to `CI` in `Mortality()`, by default "MEAN"}

\item{G}{refers to `calc_fild` in `Mortality()`}

\item{D}{refers to `conc_real` in `Mortality()`}

\item{P}{refers to `pop_total` in `Mortality()`}

\item{A}{refers to `age_struc` in `Mortality()`}

\item{M}{refers to `mort_rate` in `Mortality()`}

\item{L}{refers to `mort_lvl` in `Mortality()`}

\item{D_cf}{{data.frame}, refers to `conc_cf` in `Mortality()`}

\item{from}{scenario/year when the driving space start}

\item{to}{scenario/year when the driving space end}
}
\value{
a data.frame of decomposed attributed deaths
}
\description{
Decomposition of attributed deaths
}
\examples{
\dontrun{
  builtin_data <- list.files(system.file('extdata', package = "AttrMort"), full.names = T)

    for (f in builtin_data) {
      nm <- basename(f) |> str_extract(".+(?=\\\\.)")
      assign(nm, readxl::read_excel(f))
    }

  Decomposition(
    serie = 1,
    crf = "GEMM",
    G = grid_info,
    D = grid_exposure,
    P = grid_pop,
    A = national_age_structure,
    M = national_mortality,
    L = "location",
    from = "base2015",
    to = "SSP1-Baseline_2030"
  )
}

}
