% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GEMM_Lookup_Table}
\alias{GEMM_Lookup_Table}
\title{the GEMM C-R curves by Burnett et al.}
\format{
list
}
\usage{
GEMM_Lookup_Table
}
\description{
the GEMM C-R curves by Burnett et al.
}
\keyword{datasets}
