% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IER2010_Lookup_Table}
\alias{IER2010_Lookup_Table}
\title{the IER C-R curves by GBD2010.}
\format{
list
}
\usage{
IER2010_Lookup_Table
}
\description{
the IER C-R curves by GBD2010.
}
\keyword{datasets}
