% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MRBRT2019_Lookup_Table}
\alias{MRBRT2019_Lookup_Table}
\title{the MRBRT C-R curves for PM25 by GBD2019.}
\format{
list
}
\usage{
MRBRT2019_Lookup_Table
}
\description{
the MRBRT C-R curves for PM25 by GBD2019.
}
\keyword{datasets}
