% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NO2_CR_Lookup_Table}
\alias{NO2_CR_Lookup_Table}
\title{the C-R curves for NO2 by Larkin et al.}
\format{
list
}
\usage{
NO2_CR_Lookup_Table
}
\description{
the C-R curves for NO2 by Larkin et al.
}
\keyword{datasets}
