% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{O3_CR_Lookup_Table}
\alias{O3_CR_Lookup_Table}
\title{the C-R curves for O3 by Carey et al.}
\format{
list
}
\usage{
O3_CR_Lookup_Table
}
\description{
the C-R curves for O3 by Carey et al.
}
\keyword{datasets}
